xsk: feature uplift from libbpf
These are the commits:

libbpf: Only create rx and tx XDP rings when necessary

Prior to this commit xsk_socket__create(_shared) always attempted to
create the rx and tx rings for the socket. However this causes an
issue when the socket being setup is that which shares the fd with the
UMEM. If a previous call to this function failed with this socket
after the rings were set up, a subsequent call would always fail
because the rings are not torn down after the first call and when we
try to set them up again we encounter an error because they already
exist. Solve this by remembering whether the rings were set up by
introducing new bools to struct xsk_umem which represent the ring
setup status and using them to determine whether or not to set up the
rings.

libbpf: Restore umem state after socket create failure

If the call to xsk_socket__create fails, the user may want to retry
the socket creation using the same umem. Ensure that the umem is in
the same state on exit if the call fails by: 1. ensuring the umem
_save pointers are unmodified.  2. not unmapping the set of umem rings
that were set up with the umem during xsk_umem__create, since those
maps existed before the call to xsk_socket__create and should remain
in tact even in the event of failure.

libbpf: Ensure umem pointer is non-NULL before dereferencing

Calls to xsk_socket__create dereference the umem to access the
fill_save and comp_save pointers. Make sure the umem is non-NULL
before doing this.

Signed-off-by: Magnus Karlsson <magnus.karlsson@intel.com>